refactor(sar_io): centralize nodata detection in raster_io utilities
Move the `get_nodata_from_src` function from `src/dataset.py` to
`src/utils/raster_io.py`. This centralizes the logic for inferring
no-data values in SAR GeoTIFFs within the dedicated I/O utility module.

This improves code organization and ensures that all raster data reading
and no-data handling logic is consistently managed in one location, simplifying maintenance.
