+++ src/utils/raster_io.py
@@ -7,9 +7,27 @@ from typing import Any, Union
 
 import numpy as np
 import rasterio
+from rasterio import windows
 
-from src.dataset import get_nodata_from_src
 from src.utils.compress_sar_with_jpeg2000 import expand_from_jp2
+
+
+def get_nodata_from_src(src: rasterio.DatasetReader) -> float:
+    # return self.src.nodata if self.src.nodata is not None else -9999.0
+    if src.nodata is not None:
+        return src.nodata
+    else:
+        data = src.read(
+            1,
+            window=windows.Window(0, 0, min(128, src.width), min(128, src.height)),
+            boundless=True,
+            fill_value=0,
+        )  # Read a sample
+        if np.any(data < -9000):  # Heuristic from user
+            # If nodata is not set, assume a common nodata value for SAR data
+            return -9999.0
+        else:
+            return 0.0  # Default to 0 if not otherwise determined.
 
 
 def read_raster_data(
